AdvModelsPro/requirements.txt
django
djangorestframework
pyjwt
mysql
mysqlclient